通知再送をAPI呼び出し方式に統一しCLI実装を整理
