--- apps/fumufumu-backend/.env.notifications.example
@@ -1,12 +1,0 @@
-# Notification CLI env example
-# Copy this file to `.env.notifications` and set actual values.
-# Do not commit real values. Put production values in CI/runtime secrets.
-
-# Required
-RESEND_API_KEY=re_xxxxxxxxxxxxxxxxxxxxx
-RESEND_FROM_EMAIL="Fumufumu <no-reply@mail.fumufumu.jp>"
-
-# Optional
-APP_BASE_URL=http://localhost:3000
-RESEND_ENDPOINT=https://api.resend.com/emails
-RESEND_TIMEOUT_MS=8000
